move the closestBreadcrumbs function to a more appropriate place
